refactor: Lotto 클래스 domain 패키지로 이동
